adjust missile targets
now a missile can only target a visible enemy.
when a stealth enemy goes invisible, the missile will drift in last direction until reacquired. If the missile drifts off-screen it is deleted. If the target is destroyed, it finds a new target. If no enemies remain to be targeted the missile is deleted.